python-docker/requirements.txt
Flask==2.2.2
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.0
passlib==1.7.4